sqlite movie review(CRUD)

+++ lib/util/constant.dart
@@ -17,7 +17,7 @@ class DB {
       id INTEGER PRIMARY KEY AUTOINCREMENT,
       name TEXT,
       review TEXT,
-      rate INTEGER,
+      rate REAL
     )
 ''';
 }
